requirements.txt
flask
flask-cors
yfinance
pandas
anthropic
python-telegram-bot
schedule
requests
python-dotenv
gunicorn